Add controllers - update repository y services

+++ DiscograficaWebApi/DAL/Repository/IRepository.cs
@@ -3,7 +3,7 @@
 public interface IRepository<T> where T : class
 {
     Task<List<T>> GetAll();
-    Task<T> GetById(int id);
+    Task<T> GetById(long id);
     Task Add(T entity);
     void Edit(T entity);
     void Delete(T entity);
