+++ DiscograficaWebApi/DAL/Repository/Implements/Repository.cs
@@ -16,7 +16,7 @@ public class Repository<T> : IRepository<T> where T : class
         return await _context.Set<T>().ToListAsync();
     }
 
-    public async Task<T> GetById(int id)
+    public async Task<T> GetById(long id)
     {
         return await _context.Set<T>().FindAsync(id);
     }
